Added support for all types

+++ pantab/tests/conftest.py
@@ -107,7 +107,7 @@ def df():
             "int32": np.int32,
             "Int32": "Int32",
             "int64": np.int64,
-            "Int32": "Int32",
+            "Int64": "Int64",
             "float32": np.float32,
             "float64": np.float64,
             "bool": np.bool,
